requirements.txt
# VideoPPT服务依赖包

# Web框架
fastapi==0.119.1
uvicorn[standard]==0.38.0
jinja2==3.1.6
starlette==0.48.0

# 视频处理
opencv-python==4.12.0.88
numpy==2.2.6

# PPT生成
python-pptx==1.0.2
Pillow==12.0.0

# 视频下载
yt-dlp==2025.10.14

# HTTP请求
requests==2.32.5

# 数据验证
pydantic==1.10.24

# 其他依赖
pydub==0.25.1

